// apps/web/src/api-client/model/notification.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Waymate API
 * Carpooling backend powering the Waymate web app.
 * OpenAPI spec version: 0.1.0
 */
import type { NotificationNotificationType } from "./notificationNotificationType";
import type { NotificationPayloadProperty } from "./notificationPayloadProperty";

export interface Notification {
    /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
    id: string;
    notificationType: NotificationNotificationType;
    /** @nullable */
    referenceEntityType: string | null;
    /**
     * @nullable
     * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
     */
    referenceEntityId: string | null;
    title: string;
    body: string;
    /** @nullable */
    payload: NotificationPayloadProperty;
    /** @nullable */
    readAt: string | null;
    createdAt: string;
}